requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
dnspython==2.7.0
fastapi==0.116.1
Flask==3.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.25.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
linkify-it-py==2.0.3
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
nvidia-ml-py==12.575.51
outcome==1.3.0.post0
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
py3nvml==0.2.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyMuPDF==1.26.3
pynput==1.8.1
pynvml==12.0.0
pypandoc==1.15
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
PySocks==1.7.1
pytesseract==0.3.13
python-docx==1.2.0
python-dotenv==1.1.1
python-telegram-bot==20.7
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rsa==4.9.1
selenium==4.33.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
starlette==0.47.1
textual==3.5.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uc-micro-py==1.0.3
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.35.0
watchfiles==1.1.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
WMI==1.5.1
wsproto==1.2.0
xmltodict==0.14.2
yt-dlp==2025.6.9
